- Allianceboard: Fixed some minor (graphical) bugs - Allianceboard: Added a lot of all new icons for the categories and fixed a bug in the icon selector - Added a new class ImageUpload for better upload handling and cleanup in user settings - It is now possible to use any image as an avatar for the alliance board. It will be automatically resized to 64x64px. - Added some more general icons
git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@891 9ddb851b-07ab-4ff2-becd-2ceea2d3057d

+++ content/userconfig/general.php
@@ -12,28 +12,17 @@
           }
           elseif ($_FILES['user_avatar_file']['tmp_name']!="")
           {
-            $source=$_FILES['user_avatar_file']['tmp_name'];
-            $ims = getimagesize($source);
-            
-            //überprüft Bildgrösse
-            if ($ims[0]==BOARD_AVATAR_WIDTH && $ims[1]==BOARD_AVATAR_HEIGHT)
-            {
-                $fname = "user_".$cu->id."_".time().".gif";
-                if (file_exists(BOARD_AVATAR_DIR."/".$arr['user_avatar']))
-                    @unlink(BOARD_AVATAR_DIR."/".$arr['user_avatar']);
-                move_uploaded_file($source,BOARD_AVATAR_DIR."/".$fname);
-                if (UNIX)
-                	chmod(BOARD_AVATAR_DIR."/".$fname,FILE_UPLOAD_PERMS);
-                
-								$cu->avatar = $fname;
-                echo "Eigenen Avatar gespeichert!<br/>";
-            }
-            else
-            {
-                echo "Fehler! Das Avatarbild hat die falsche Gr&ouml;sse!<br/>";
-            }
-          }
-          
+          	$imup = new ImageUpload('user_avatar_file',BOARD_AVATAR_DIR,"user_".$cu->id."_".time());
+          	$imup->setMaxSize(BOARD_AVATAR_MAX_SIZE);
+          	$imup->setMaxDim(BOARD_AVATAR_MAX_WIDTH,BOARD_AVATAR_MAX_HEIGHT);
+          	$imup->enableResizing(BOARD_AVATAR_WIDTH,BOARD_AVATAR_HEIGHT);
+          	
+						if ($imup->process())
+						{
+							$cu->avatar = $imup->getResultName();
+							ok_msg("Avatar hochgeladen!");
+						}           	
+					}
           
           // Profil-Bild
           if (isset($_POST['profile_img_del']) && $_POST['profile_img_del']==1)
@@ -42,61 +31,18 @@
           }
           elseif ($_FILES['user_profile_img_file']['tmp_name']!="")
           {
-          	if ($_FILES['user_profile_img_file']['size']<=PROFILE_IMG_MAX_SIZE)
-          	{          		
-              $source=$_FILES['user_profile_img_file']['tmp_name'];
-              $ims = getimagesize($source);
-              
-             	$ext = substr($ims['mime'],strrpos($ims['mime'],"/")+1);
-             	if ($ext=="jpg" || $ext=="jpeg" || $ext=="gif" || $ext=="png")
-             	{                  
-                //überprüft Bildgrösse
-                if ($ims[0]<=PROFILE_MAX_IMG_WIDTH && $ims[1]<=PROFILE_MAX_IMG_HEIGHT)
-                {
-                    $fname = "user_".$cu->id."_".time().".".$ext;
-                    if (file_exists(PROFILE_IMG_DIR."/".$arr['user_profile_img']))
-                        @unlink(PROFILE_IMG_DIR."/".$arr['user_profile_img']);
-                    move_uploaded_file($source,PROFILE_IMG_DIR."/".$fname);
-		                if (UNIX)
-		                	chmod(PROFILE_IMG_DIR."/".$fname,FILE_UPLOAD_PERMS);
-		                	
-                    if ($ims[0]>PROFILE_IMG_WIDTH || $ims[1]>PROFILE_IMG_HEIGHT)
-										{
-											if (resizeImage(PROFILE_IMG_DIR."/".$fname,PROFILE_IMG_DIR."/".$fname,PROFILE_IMG_WIDTH,PROFILE_IMG_HEIGHT,$ext))
-											{
-												echo "Bildgrösse wurde angepasst! ";
-                      	echo "Profilbild gespeichert!<br/>";
-	                    	$cu->profileImage = $fname;
-											}
-											else
-											{
-												Echo "Bildgrösse konnte nicht angepasst werden!";
-                        @unlink(PROFILE_IMG_DIR."/".$arr['user_profile_img']);
-											}
-										}
-										else
-										{
-                    	echo "Profilbild gespeichert!<br/>";
-                    	$cu->profileImage = $fname;
-                    }
-                }
-                else
-                {
-                    echo "Fehler! Das Profilbild hat die falsche Gr&ouml;sse (".$ims[0]."*".$ims[1].")!<br/>";
-                }
-             	}
-             	else
-             	{
-                echo "Fehler! Das Profilbild muss vom Typ jpeg, png oder gif sein.!<br/>";
-							}
-						}	                 	
-           	else
-           	{
-              echo "Fehler! Das Profilbild ist zu gross (Max ".nf(PROFILE_IMG_MAX_SIZE)." Byte)!<br/>";
-						}
+          	$imup = new ImageUpload('user_profile_img_file',PROFILE_IMG_DIR,"user_".$cu->id."_".time());
+          	$imup->setMaxSize(PROFILE_IMG_MAX_SIZE);
+          	$imup->setMaxDim(PROFILE_MAX_IMG_WIDTH,PROFILE_MAX_IMG_HEIGHT);
+          	$imup->enableResizing(PROFILE_IMG_WIDTH,PROFILE_IMG_HEIGHT);
+          	
+						if ($imup->process())
+						{
+							$cu->profileImage = $imup->getResultName();
+							ok_msg("Profilbild hochgeladen!");
+						}          	
           }
-          
-          
+
           $cu->email = $_POST['user_email'];
           $cu->profileText = addslashes($_POST['user_profile_text']);
           $cu->signature = addslashes($_POST['user_signature']);
@@ -143,8 +89,12 @@
       	<td>";
         if ($cu->profileImage!="")
         {
-          echo '<img src="'.PROFILE_IMG_DIR.'/'.$cu->profileImage.'" alt="Profil" /><br/>';
-          echo "<input type=\"checkbox\" value=\"1\" name=\"profile_img_del\"> Bild l&ouml;schen<br/>";
+        	$im = PROFILE_IMG_DIR.'/'.$cu->profileImage;
+        	if (is_file($im))
+        	{
+	          echo '<img src="'.$im.'" alt="Profil" /><br/>';
+	          echo "<input type=\"checkbox\" value=\"1\" name=\"profile_img_del\"> Bild l&ouml;schen<br/>";
+        	}
         }
       	echo "Profilbild heraufladen/&auml;ndern: <input type=\"file\" name=\"user_profile_img_file\" /><br/>
       	<b>Regeln:</b> Max ".PROFILE_MAX_IMG_WIDTH."*".PROFILE_MAX_IMG_HEIGHT." Pixel, Bilder grösser als 
@@ -166,9 +116,9 @@
       	echo "Eigener Avatar heraufladen/&auml;ndern (".BOARD_AVATAR_WIDTH."*".BOARD_AVATAR_HEIGHT." Pixel, GIF): <input type=\"file\" name=\"user_avatar_file\" /></td>
       </tr>";
       echo "<tr>
-      	<th width=\"35%\">Öffentliches Foren-Profil:<br/>
-      	<span style=\"font-weight:500;font-size:7pt;\">(zb http://www.etoa.ch/forum/profile.php?userid=1)</span></th>
-      	<td width=\"65%\"><input type=\"text\" name=\"user_profile_board_url\" maxlength=\"200\" size=\"50\" value=\"".$cu->profileBoardUrl."\"></td>
+      	<th width=\"35%\">Öffentliches Foren-Profil:</th>
+      	<td width=\"65%\"><input type=\"text\" name=\"user_profile_board_url\" maxlength=\"200\" size=\"50\" value=\"".$cu->profileBoardUrl."\"><br/>
+      	<span style=\"font-weight:500;font-size:7pt;\">(zb http://www.etoa.ch/forum/profile.php?userid=1)</span></td>
       </tr>";
 
       tableEnd();
